fix: circular references
